lib: shared memory map
The shared memory map will allow passing SDU's between IPCPs
efficiently. It is a single block of managed memory. SDU's are stored
in a shm_du_buff datastructure, which is a list of blocks inside the
shared memory region. The mechanism for passing SDU's is not yet
implemented and could be either within the region or using another
local IPC mechanism.

The following things are configured at compile time:

SHM_FILE_NAME         : the name. Default is "ouroboros_du_map".
SHM_DU_BUFF_BLOCK_SIZE: size of a block inside the shared memory
                        region. Default is the pagesize of the system.
SHM_DU_MAP_SIZE       : the approximate size in bytes of the shared
                        memory map. The actual filesize may differ by
                        a small margin. Default is 4MB.