lib: Shared Memory
Updated tests. Added code to drop a corner case (packet where the tail
PCI would cross the edge of the ring buffer) because solving this very
rare case is not worth the performance hit on the ringbuffer the extra
code would incur. This means the ringbuffer might drop a very small
percentage of packets.