shim-udp: flow allocation added
This implements the API for flow allocation. The shims currently does
the following.

The shim IPCP binds to an interface (IP address) and listens for flow
allocation requests on UDP port 0x0D1F (3359), referenced as the
listen port (lp). It will treat any datagram received on lp as a flow
allocation request.

Upon receiving an allocation request IRM_MSG_CODE__IPCP_FLOW_ALLOC
from the IRMd, the shim IPCP will bind a UDP socket to a port (cp)
allocated by the host OS. From that port it will send a UDP packet
containing the destination ap_name to server_host:lp and wait for a
response.

Upon reception of a packet on server_host:lp, the shim_IPCP creates a
UDP socket for the flow with a port set by the host os (sp), binds to
it and echoes the received datagram back from server_host:sp to
client_host:cp. It will also notify the IRMd of an incoming flow
allocation request IRM_MSG_CODE__IPCP_FLOW_REQ_ARR, with as
src_ap_name ("John Day").  It will get the port_id as a return value
of that message and create a flow with status FLOW_PENDING with that
port_id. If the server responds negatively to the flow allocation
request (i.e. the shim IPCP on the server side receives a
IRM_MSG_CODE__IPCP_FLOW_ALLOC_RESPONSE with a response != 0, it will
delete the pending flow. If response == 0, it will set the status to
FLOW_ALLOCATED.

On the client machine the IPCP will learn sp upon reception of the
echoed datagram. It will then create a flow with the port_id it
received with the message from the IRMd and set it to ALLOCATED.

Pending implementation:

DNS support, this PR only supports local flows on the loopback adapter
127.0.0.1.

A thread to listen for the echoed message, to avoid the
entire IPCP to block when the echoed message is lost.

This PR compiles but is untested pending necessary implementations
elsewhere in the stack.