lib, ipcpd, irmd: Add QoS cube definition
This adds the QoS cube definition, which is an enum to select which
QoS is needed in the IPCP. An application has to use the qos_spec in
qos.h to define what it needs. The IRMd will map this unto a qos cube
definition.

Some headers are now also no longer installed on the system, since
they are only to be used within the irmd and ipcps.