ipcpd: shim-udp: Revised locking
Simplified locking to take only two locks: the first lock guards the
state of the ipcp. This lock must be held for writing on bootstrap and
closing, and held for reading during all other operations. The second
lock guards operations on flows, and must be held for writing during
allocation and deallocation, and held for reading when sending sdu's.
After adding a fd to FD_SET, the shim will wait for 1 ms to ensure
that the FD is added to the select call.