ipcpd: shim-ipcp: fixed race conditions
There were two race conditions. The first was where client received
the response message before the client UDP port was connected to the
server UDP port and the first message of the client was sent to the
LISTEN port instead of the server application port. This was solved by
reordering the operations so the response is sent after the port is
correctly connected. The second race condition is that the FD is not
added to the set in time at the server side before the first SDU
arrives. This was solved by adding a variable proteceted by a mutex.