irmd, lib: use shm_du_map as lockfile.
Added the pid of the irmd to the shm_du_map. The IRMd will check for
an existing shm_du_map. If there is an existing file, it will exit if
the owner IRMd is running or remove it if the owner IRMd is not
running.

Also simplifies calculation of the shm_du_map pointers and corrects
exiting calls for the IRMd.

Fixes #8.