ipcpd: Adds RX_RING and TX_RING in shim-eth-llc
This adds the RX_RING and TX_RING socket options to the shim-eth-llc
for faster transmission and reception. On systems where these options
are not available the old method of sending and receiving is
used. Further optimization would be creating and assigning a
read/write thread per CPU core, but that becomes very platform
specific, and is not required in the short term.

Fixes #10