irmd: Revised flow allocation
Flow allocation requests and registered api states revised so all
states are tracked with a condition variable. This is a more reliable
approach and improves stability of flow allocation.

Some other refactoring was also done, such as renaming port_map_entry
to irm_flow and hiding some internal structures of the registry.