ipcpd: shim-eth-llc: Use correct frame length
Upon receipt of a frame, the minimum of the length reported in the
frame and the frame length reported by kernel is taken. Some device
drivers change the length in the frame. Some others add padding,
making the length reported by the kernel too high. This is a lousy
workaround to make it work on as many systems as possible.