ipcpd: Revise PCI manipulation in normal
The pci function will store offsets to avoid having to recalculate
them every time. TTL handling is removed from the fmgr, the TTL is
automatically decreased when deserializing the PCI (or set to 1 if
there is no TTL present so the PDU will be forwarded). The
deserialisation function now takes a pointer to a struct pci as input
to avoid memory allocation and release each time a PDU is
processed. Some checks have been replaced with assert() and return
values replaced with void where it makes sense.