lib, tools, ipcpd: Fix compilation on FreeBSD
FreeBSD defines its own bswap64 in the <sys/endian.h> header, which is
now included correctly.

POSIX requires XSI or RTS extensions to be enabled for the
sa_sigaction field to be visible. Linux doesn't do the check
correctly, but FreeBSD does.

Removes the LOG_MISSING call which was deprecated with the last
revision of the logging system.