include: Add header for endianness
This adds a header for dealing with endianness in ouroboros. It is
extracted from the byte_order header in the library (which now
includes this header). It also exposes the functions ntohl, ntohll,
htonl and htonll, necessary for converting 32 and 64 bit values for
storage and retrieval from the RIB (which should store multi-byte
values in network byte order).