lib: Exchange protocol info during CACEP
This exchanges a protocol name, a protocol version and concrete syntax
for the protocol upon CACEP. For CDAP, only version 1 and GPB are
supported. No lists for other supported versions or syntaxes are
exchanged (but the proto file supports it). CACEP fails if there is a
mismatch between the protocol names, version and syntax specified by
the communicating parties.