irmd: Don't call ipcp_flow_dealloc under lock
When a process was killed, ipcp_flow_dealloc was called under a lock,
causing a deadlock in the IRMd because handling the IPCP response also
needs to take the same lock (the IPCP calls flow_dealloc to finalize
this). This deadlock also resulted in the IPCP always reporting that
it failed to send a reply message as the deadlock effectively blocks
the IRMd until its socket timeout expired and thus the IPCP was always
responding to an already closed socket.