ipcpd: Fix deadlock in DHT
The lookup_update was waiting for the LU_INIT state to resolve under
dht->lock which prevented that update.

This PR fixes this bug, but leaves a (very rare) bug when the
lookup_destroy is called while the lookup_update is waiting for the
LU_INIT state to resolve. The solution also is a (harmless) lock
inversion, but this is also not the best.

Fixes #51
Fixes #52