ipcpd: Handle cancellation while writing LSAs
This adds a cleanup handler to ensure the lock is released in the
event that the thread is cancelled during a write. An alternative
solution could be to make the flow write in a non-blocking way (only
the blocking write contains an execution path that has a cancellation
point).

Fixes #55