tools: Increase read buffer in ocbr server
The ocbr server had a 1500 byte read buffer, which caused it to
perform partial reads on larger SDUs, slowing it down considerably
when sendin large packets. The buffer has been increased to 512KB,
partial reads disabled, and the client will give an error when larger
frames are sent. It will also warn if the size overflows (avoiding a
SEGV).

Signed-off-by: Dimitri Staessens <dimitri.staessens@ugent.be>
Signed-off-by: Sander Vrijders <sander.vrijders@ugent.be>