ipcpd: Fix bad memory handling in LFA policy
The Loop-Free Alternates policy had bad memory management in two
places. In the calculation of the LFAs a table was freed in the first
iteration of a loop, whereas it was still needed in the other
iterations. It is now freed outside of the loop. In the alternate PFF
the address structs were not freed upon shutdown, this has been added
as well. It also fixes some bad initialization in the LFA calculation
function.

Signed-off-by: Sander Vrijders <sander.vrijders@ugent.be>
Signed-off-by: Dimitri Staessens <dimitri.staessens@ugent.be>